backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pandas
numpy
scikit-learn
joblib
pulp
supabase
python-dotenv
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
fastapi
uvicorn
httpx
